+++ src/utils/Main/url.ts
@@ -7,10 +7,12 @@ export const panelUrlDev = `http://localhost:${process.env.DEV_PANEL_PORT ?? "51
 export const settingsUrlDev = `http://localhost:${process.env.DEV_SETTINGS_PORT ?? "5173/settings.html"}`;
 
 export const changelogUrlDev = `http://localhost:${process.env.DEV_SETTINGS_PORT ?? "5173/changelog.html"}`;
+export const previewUrlDev = `http://localhost:${process.env.DEV_SETTINGS_PORT ?? "5173/preview.html"}`;
 
 export const panelUrlProd = `file://${resolve(__dirname, "../index.html")}`;
 export const settingsUrlProd = `file://${resolve(__dirname, "../settings.html")}`;
 export const changelogUrlProd = `file://${resolve(__dirname, "../changelog.html")}`;
+export const previewUrlProd = `file://${resolve(__dirname, "../preview.html")}`;
 
 export const preloadMainScriptPathDev = `${resolve(
   __dirname,